Fix spirit reaper
If it was being targeted by multiple cards in the same chain, and the last card in the chain was negated, it woudln't have been destroyed after resolving the other cards in the chain targeting it. Also shouldn't self destruct if targeted during the damage calculation, but should self destroy at teh end of the battle like if targeted during the damage step.