Fix for "Orichalcos Kyutora" (#448)
* Fix for "Orichalcos Kyutora"

Damage was only applying to Kyutora and not to other monsters controlled by the player (e4 was also contributing to the issue, and was extraneous since e2 reduced the damage to 0 anyway). Also changed it so that OCG Shunoros could not be SS'd by this effect (OCG Shunoros requires a Normal monster to be destroyed to be SS'd).

https://github.com/ProjectIgnis/CardScripts/issues/185 - In response to this issue